pyproject.toml
[tool.poetry]
name = "copilot-prototype"
version = "0.1.0"
description = ""
authors = ["vincent-goldberg <vincent.goldberg@berkeley.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"

langchain = "^0.0.300"
streamlit = "^1.27.0"
unstructured="*"
nltk="*"
black = "*"
pinecone-client = "*"
fastapi = "*"
jinja2 = "*"
uvicorn = "*"
streamlit-chat = "*"
tqdm = "*"
pypdf = "^3.16.1"
sentence-transformers = "^2.2.2"
watchdog = "^3.0.0"
pypdf2 = "^3.0.1"
faiss-cpu = "^1.7.4"
chromadb = "^0.4.15"
python-dotenv = "^1.0.0"
tiktoken = "^0.5.1"
langchainhub = "^0.1.13"
runpod = "^0.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
